+++ packages/table-render/src/hooks.ts
@@ -1,9 +1,11 @@
 import { useReducer, useContext } from 'react';
-import { Ctx, StoreCtx } from './context';
+import { Ctx, StoreCtx } from './Context';
 import { TableContext } from './interface';
 
 // 使用最顶层组件的 setState
-export const useTable: <RecordType extends object = any>() => TableContext<RecordType> = () => {
+export const useTable: <RecordType extends object = any>() => TableContext<
+  RecordType
+> = () => {
   return useContext(Ctx);
 };