+++ tools/schema-generator/src/Provider.jsx
@@ -57,8 +57,8 @@ function Provider(props, ref) {
     to: schema => schema,
     fromSetting,
     toSetting,
-    ..._transformer
-  }
+    ..._transformer,
+  };
 
   const frwRef = ref || useRef();
   const [state, setState] = useSet({
@@ -69,22 +69,17 @@ function Provider(props, ref) {
     schema: {},
     selected: undefined, // 被选中的$id, 如果object/array的内部，以首字母0标识
   });
-  const [itemError, setItemError] = useState([])
+  const [itemError, setItemError] = useState([]);
 
   // 收口点 propsSchema 到 schema 的转换 (一共3处，其他两个是 importSchema 和 setValue，在 FRWrapper 文件)
   useEffect(() => {
-    const schema = defaultValue ? transformer.from(defaultValue) : DEFAULT_SCHEMA;
+    const schema = defaultValue
+      ? transformer.from(defaultValue)
+      : DEFAULT_SCHEMA;
     if (schema) setState(schemaToState(schema));
   }, [defaultValue]);
 
-  const {
-    formData,
-    frProps,
-    isNewVersion,
-    preview,
-    schema,
-    selected,
-  } = state;
+  const { formData, frProps, isNewVersion, preview, schema, selected } = state;
 
   const onChange = data => {
     setState({ formData: data });