+++ tools/schema-generator/src/index.js
@@ -5,22 +5,19 @@ import Canvas from './components/Canvas';
 import Settings from './components/Settings';
 import './styles/index.less';
 
-const Generator = forwardRef(({
-  fixedName,
-  settingsWidgets,
-  onCanvasSelect,
-  ...props
-}, ref) => {
-  return (
-    <Provider ref={ref} {...props}>
-      <div className="fr-generator-container">
-        <Sidebar fixedName={fixedName} />
-        <Canvas onSelect={onCanvasSelect} />
-        <Settings widgets={settingsWidgets} />
-      </div>
-    </Provider>
-  );
-});
+const Generator = forwardRef(
+  ({ fixedName, settingsWidgets, onCanvasSelect, ...props }, ref) => {
+    return (
+      <Provider ref={ref} {...props}>
+        <div className="fr-generator-container">
+          <Sidebar fixedName={fixedName} />
+          <Canvas onSelect={onCanvasSelect} />
+          <Settings widgets={settingsWidgets} />
+        </div>
+      </Provider>
+    );
+  }
+);
 
 Generator.Provider = Provider;
 Generator.Sidebar = Sidebar;
@@ -32,8 +29,5 @@ export {
   defaultCommonSettings,
   defaultGlobalSettings,
 } from './settings';
-export {
-  fromSetting,
-  toSetting,
-} from './transformer/form-render';
+export { fromSetting, toSetting } from './transformer/form-render';
 export default Generator;