+++ docs/table-render/demo/noTitle.tsx
@@ -49,12 +49,12 @@ const Demo = () => {
         }
       })
       .catch(e => {
-        console.log('Oops, error', e)
+        console.log('Oops, error', e);
         // 注意一定要返回 rows 和 total
         return {
           rows: [],
           total: 0,
-        }
+        };
       });
   };
 
@@ -137,10 +137,7 @@ const Demo = () => {
         afterSearch={params => console.log('afterSearch', params)}
         api={searchApi}
       />
-      <Table
-        columns={columns}
-        rowKey="id"
-      />
+      <Table columns={columns} rowKey="id" />
     </div>
   );
 };