+++ tools/schema-generator/src/index.d.ts
@@ -1,4 +1,4 @@
-import type { FC, ReactNode, ForwardRefExoticComponent, RefAttributes } from 'react';
+import { FC, ReactNode, ForwardRefExoticComponent, RefAttributes } from 'react';
 
 export interface Transformer {
   /** 正向的转换函数 */
@@ -76,8 +76,18 @@ export interface FRGeneratorProps {
   settingsWidgets?: any;
   /** 组件和 schema 的映射规则 */
   mapping?: any;
-  fieldRender?: (schema?: any, widgetProps?: any, children?: ReactNode, originNode?: ReactNode) => ReactNode;
-  fieldWrapperRender?: (schema?: any, isSelected?: boolean, children?: ReactNode, originNode?: ReactNode) => ReactNode;
+  fieldRender?: (
+    schema?: any,
+    widgetProps?: any,
+    children?: ReactNode,
+    originNode?: ReactNode
+  ) => ReactNode;
+  fieldWrapperRender?: (
+    schema?: any,
+    isSelected?: boolean,
+    children?: ReactNode,
+    originNode?: ReactNode
+  ) => ReactNode;
   /** 表单 data 变化回调 */
   onChange?: (data: any) => void;
   /** 表单 schema 变化回调 */